Fixed the issue, added gitignore
CBODYAsteroid is a static variable but you assign it in the constructor. That means every time you create an asteroid you make a new CBType and it is not the same as the previous asteroid when nape compares them. You should assign it at the top like so public static var CBODYAsteroid:CbType = new CbType(); instead of in new.

The other thing is you add the cbType in new but then in create in the asteroid you call createCircularBody which creates a new body and won't have the cbType you added in new. So you can either do everything in new or add the cbType in create after you make the body. The same goes for setting the userdata.
Credit goes to @codescapade